requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
itsdangerous==0.23
psycopg2==2.5.2
python-dateutil==2.2
six==1.5.2
gunicorn==0.17.2